requirements.txt
regex
ipykernel
nltk==3.9.1
numpy==1.24.4
pandas
scikit-learn
gensim
transformers
datasets
# Project1
diffusers
accelerate
torch
matplotlib
# Project2
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
evaluate
# Telegram
aiogram==2.25.1
openai==0.28
python-dotenv
# Image Dalle
flask
langchain
llama-index
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain_community
huggingface_hub
numexpr
bitsandbytes
sentence-transformers
tiktoken
faiss-cpu
chromadb==0.4.15
weaviate-client
unstructured
unstructured[pdf]
pymupdf
transformers
einops
accelerate 
langchain-google-genai